requirements.txt
viam-sdk
adafruit_dht
adafruit-circuitpython-dht
board